Issue #356: Fixing stb.h in bootloader for Windows-64bit. Some fixes for Linux as well.

+++ bootloader/common/pyi_python.h
@@ -105,7 +105,7 @@ EXTDECLPROC(int, PyObject_SetAttrString, (PyObject *, char *, PyObject *));
 EXTDECLPROC(PyObject *, PyList_New, (int));
 EXTDECLPROC(int, PyList_Append, (PyObject *, PyObject *));
 EXTDECLPROC(PyObject *, Py_BuildValue, (char *, ...));
-EXTDECLPROC(PyObject *, PyString_FromStringAndSize, (const char *, int));
+EXTDECLPROC(PyObject *, PyString_FromStringAndSize, (const char *, size_t));
 EXTDECLPROC(PyObject *, PyFile_FromString, (char *, char *));
 EXTDECLPROC(char *, PyString_AsString, (PyObject *));
 EXTDECLPROC(PyObject *, PyObject_CallFunction, (PyObject *, char *, ...));
